requirements.txt
fastapi
uvicorn
python-multipart
tensorflow-cpu
numpy
scikit-image
trimesh